requirements.txt
# Django Core
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0

# JWT Authentication
djangorestframework-simplejwt>=5.3.0

# Environment Variables
python-decouple>=3.8

# Excel File Handling
openpyxl>=3.1.0

# Database (PostgreSQL support - optional)
psycopg2-binary>=2.9.0

# API Documentation
drf-yasg>=1.21.7

# Code Quality
flake8>=6.1.0
black>=23.0.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0




